+++ src/components/@common/Footer/index.jsx
@@ -2,20 +2,20 @@ const Footer = () => {
   
   return (
     <div className="bg-white h-52 pl-24 flex content-center max-w-screen-xl">
-      <div className="content-center">
-        <h2 className="mb-2">북 스포일러</h2>
-        <p className="text-sm gray5">
+      <div className="content-center gray5">
+        <h2 className="mb-2 font14">북 스포일러</h2>
+        <p className="font12 gray5">
           상호명 : 주식회사 굿플로우｜사업자등록번호 : 793-87-02421｜대표
           :김민영
         </p>
-        <p className="text-sm gray5">통신판매업 : 2023-경기파주-1478</p>
-        <p className="text-sm gray5">
+        <p className="font12 gray5">통신판매업 : 2023-경기파주-1478</p>
+        <p className="font12 gray5">
           주소 : 경기도 파주시 교하로 159번길 33, 목동프라자 에이동 397호
           (목동동)
         </p>
-        <p className="text-sm gray5">개인정보관리책임자 : 김장미</p>
-        <p className="text-sm gray5">문의 : activejjjm@naver.com</p>
-        <p className="text-sm gray5">
+        <p className="font12 gray5">개인정보관리책임자 : 김장미</p>
+        <p className="font12 gray5">문의 : activejjjm@naver.com</p>
+        <p className="font12 gray5">
           Copyright ⓒ 2024 BOOKSPOILER. All rights reserved.
         </p>
       </div>
