+++ src/pages/BookSpoiler/BookSpoilerDetail.jsx
