+++ src/pages/index.js
@@ -1,19 +1,19 @@
-export { default as AdminBookCreatePage } from './AdminBookCreatePage';
-export { default as AdminBookListPage } from './AdminBookListPage';
-export { default as AdminChellengePage } from './AdminChellengePage';
-export { default as AdminCommentPage } from './AdminCommentPage';
-export { default as AdminUserPage } from './AdminUserPage';
-export { default as BookSpoilerDetail } from './BookSpoilerDetail';
-export { default as BookSpoilerPage } from './BookSpoilerPage';
-export { default as ChallengeActionPage } from './ChallengeActionPage';
-export { default as ChallengePage } from './ChallengePage';
-export { default as IndexPage } from './IndexPage';
-export { default as LoginPage } from './LoginPage';
-export { default as MyBookSpoilerPage } from './MyBookSpoilerPage';
-export { default as MyChallengePage } from './MyChallengePage';
-export { default as MyInfoPage } from './MyInfoPage';
-export { default as MyPaymentListPage } from './MyPaymentListPage';
-export { default as PaymentPage } from './PaymentPage';
+export { default as AdminBookCreatePage } from './Admin/AdminBookCreatePage';
+export { default as AdminBookListPage } from './Admin/AdminBookListPage';
+export { default as AdminChellengePage } from './Admin/AdminChellengePage';
+export { default as AdminCommentPage } from './Admin/AdminCommentPage';
+export { default as AdminUserPage } from './Admin/AdminUserPage';
+export { default as BookSpoilerDetail } from './BookSpoiler/BookSpoilerDetail';
+export { default as BookSpoilerPage } from './BookSpoiler/BookSpoilerPage';
+export { default as ChallengeActionPage } from './Challenge/ChallengeActionPage';
+export { default as ChallengePage } from './Challenge/ChallengePage';
+export { default as IndexPage } from './Home/IndexPage';
+export { default as LoginPage } from './Login/LoginPage';
+export { default as MyBookSpoilerPage } from './MyPage/MyBookSpoilerPage';
+export { default as MyChallengePage } from './MyPage/MyChallengePage';
+export { default as MyInfoPage } from './MyPage/MyInfoPage';
+export { default as MyPaymentListPage } from './MyPage/MyPaymentListPage';
+export { default as PaymentPage } from './Payment/PaymentPage';
 export { default as StyleGuide } from './ui/StyleGuide';
